// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          avatar_url: string | null
          bio: string | null
          child_personal_state: string | null
          child_professional_state: string | null
          course_details: string | null
          created_at: string
          display_name: string | null
          email: string | null
          experience: string | null
          expertise: string | null
          id: string
          location: string | null
          personal_emotions: string | null
          phone: string | null
          qualifications: string | null
          student_email: string | null
          updated_at: string
          user_type: Database["public"]["Enums"]["user_type"] | null
        }
        Insert: {
          avatar_url?: string | null
          bio?: string | null
          child_personal_state?: string | null
          child_professional_state?: string | null
          course_details?: string | null
          created_at?: string
          display_name?: string | null
          email?: string | null
          experience?: string | null
          expertise?: string | null
          id: string
          location?: string | null
          personal_emotions?: string | null
          phone?: string | null
          qualifications?: string | null
          student_email?: string | null
          updated_at?: string
          user_type?: Database["public"]["Enums"]["user_type"] | null
        }
        Update: {
          avatar_url?: string | null
          bio?: string | null
          child_personal_state?: string | null
          child_professional_state?: string | null
          course_details?: string | null
          created_at?: string
          display_name?: string | null
          email?: string | null
          experience?: string | null
          expertise?: string | null
          id?: string
          location?: string | null
          personal_emotions?: string | null
          phone?: string | null
          qualifications?: string | null
          student_email?: string | null
          updated_at?: string
          user_type?: Database["public"]["Enums"]["user_type"] | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      user_type: "student" | "parent" | "therapist"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never